Fixed issue with fastdata assisted threading params to use uint16_t instead of uint8_t as this was causing issues when communicating with the firmware; Started fixing bugs with scales not always having positive ratios; Fixed issue with retraction jog_speed; Fixed issue with servo not being turned off when stopping Assisted Threading; Fixed issue with go_to_start step not properly checking that cross_slide has been retracted;

+++ rcp/utils/devices.py
@@ -66,11 +66,11 @@ typedef struct {
   uint32_t executionInterval;
   uint16_t servoEnable;
   uint32_t threadDesiredSteps;
-  uint8_t threadRequest;
-  uint8_t threadReset;
+  uint16_t threadRequest;
+  uint16_t threadReset;
   uint16_t threadSpindleIndex;
   uint32_t threadPhaseRef;
-  uint8_t threadHasPhase;
+  uint16_t threadHasPhase;
   uint16_t threadEnabled;
   uint16_t threadTolerance;
 } fastData_t;
